Use `CMAKE_CURRENT_SOURCE_DIR` instead of `CMAKE_SOURCE_DIR`
This fixes the issue of CMake not being able to find
the `SetSystemProcessor` module.

`CMAKE_CURRENT_SOURCE_DIR` points to the current `CMakeLists.txt`
being processed, which in this case is in the root of the repo.

`CMAKE_SOURCE_DIR` points to the top-level root of the project,
which is not the same as soxr's root directory when the library
is used as an external project.

This is a problem when using `soxr` as an external project
using CMake's `ExternalProject` module. Since in that case
CMake is unable to find the `SetSystemProcessor` (included
as `${CMAKE_SOURCE_DIR}/cmake/Modules`) because the variable
is expanded to the outer project's root, not `soxr`.